refactor(authz): rlss editor ui
