+++ apps/frontend/src/lib/authz/rls/RLSItem.svelte
@@ -25,7 +25,6 @@
 </script>
 
 <li class="flex items-start gap-2">
-	<Select class="w-30" items={actions} bind:value={rls.policy.action} />
 	<div class="flex-1 w-full p-1 rounded-md bg-gray-100 border border-gray-200">
 		<FilterEditor bind:value={rls.policy.filter} let:add>
 			<Button on:click={add} class="w-full mt-2" size="xs" color="alternative">{$t('Create New Filter')}</Button>
