+++ apps/frontend/src/lib/authz/rls/RLSModal.svelte
@@ -5,6 +5,6 @@
 	import { t } from '$lib/i18n'
 </script>
 
-<Modal bind:open={$rlsModal.open} class="w-full" title={$t('rls', { ns: 'authz' })} placement="top-center" size="lg">
+<Modal bind:open={$rlsModal.open} class="w-full" title={$t('rls', { ns: 'authz' })} placement="top-center" size="xl">
 	<RLS />
 </Modal>
