+++ apps/frontend/src/lib/authz/rls/RLSTabs.svelte
@@ -1,0 +1,11 @@
+<script lang="ts">
+	import { Tabs } from 'flowbite-svelte'
+	import { rlsActions } from '@undb/authz'
+	import RlsTab from './RLSTab.svelte'
+</script>
+
+<Tabs style="underline">
+	{#each rlsActions as action}
+		<RlsTab {action} />
+	{/each}
+</Tabs>
